fix: preserve full DSP cycle counts

+++ dsp/src/kernel/knl_profile.h
@@ -59,7 +59,7 @@ extern "C" {
 
 typedef struct {
     uint32_t period;
-    uint32_t cycles;
+    uint64_t cycles;
 
 } t_profile;
